requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-multipart==0.0.20
jinja2==3.1.5
google-genai==0.7.0
Pillow==10.3.0
